requirements.txt
awscli==1.25.50
boto3==1.24.50
botocore==1.27.50
click==8.1.3
colorama==0.4.4
docutils==0.16
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
